backend/app/services/ai_analysis.py
import json, re
from pathlib import Path
from app.config import settings

try:
    from openai import OpenAI
    _openai_ok = True
except ImportError:
    _openai_ok = False

try:
    import pytesseract
    from PIL import Image
    _ocr_ok = True
except ImportError:
    _ocr_ok = False

try:
    import fitz
    _pdf_ok = True
except ImportError:
    _pdf_ok = False

SYSTEM_PROMPT = """You are MediClear AI, an expert medical report analyzer.
Analyze lab values and return ONLY valid JSON (no markdown, no backticks):
{
  "reportType": "Test Name",
  "values": {
    "Test Name": {"val": 14.2, "unit": "g/dL", "min": 13.5, "max": 17.5, "status": "normal"}
  },
  "summary": "2-3 sentence professional summary",
  "explanation": "2-3 sentence patient-friendly analogy-based explanation",
  "tips": ["tip1", "tip2", "tip3", "tip4"]
}
status must be exactly: normal | high | low"""

def extract_text(file_path: str) -> str:
    ext = Path(file_path).suffix.lower()
    if ext == ".pdf" and _pdf_ok:
        try:
            doc = fitz.open(file_path)
            text = "".join(p.get_text() for p in doc)
            doc.close()
            return text.strip()
        except: pass
    if ext in [".jpg", ".jpeg", ".png", ".bmp"] and _ocr_ok:
        try:
            return pytesseract.image_to_string(Image.open(file_path)).strip()
        except: pass
    return ""

def call_openai(prompt: str) -> dict:
    client = OpenAI(api_key=settings.OPENAI_API_KEY)
    r = client.chat.completions.create(
        model="gpt-4o",
        messages=[{"role":"system","content":SYSTEM_PROMPT},{"role":"user","content":prompt}],
        temperature=0.3, max_tokens=1200
    )
    raw = re.sub(r"```json|```","",r.choices[0].message.content.strip()).strip()
    return json.loads(raw)

def demo_analysis(filename: str) -> dict:
    fn = filename.lower()
    if any(k in fn for k in ["lipid","chol"]):
        return {"reportType":"Lipid Panel","values":{"Total Cholesterol":{"val":235,"unit":"mg/dL","min":0,"max":200,"status":"high"},"LDL":{"val":158,"unit":"mg/dL","min":0,"max":100,"status":"high"},"HDL":{"val":38,"unit":"mg/dL","min":40,"max":999,"status":"low"},"Triglycerides":{"val":210,"unit":"mg/dL","min":0,"max":150,"status":"high"}},"summary":"Lipid panel reveals significantly elevated LDL and total cholesterol with reduced HDL, indicating high cardiovascular risk. Triglycerides are also elevated, compounding the atherosclerotic risk profile.","explanation":"Imagine your arteries as water pipes. LDL ('bad') cholesterol is like rust building up inside — yours is higher than ideal. HDL acts like a cleaning crew — yours is a bit low. Together, your pipes need some care to stay clear and healthy long-term.","tips":["Reduce saturated fats: limit red meat, butter, full-fat dairy","Increase soluble fiber: oats, apples, beans every day","30 minutes of cardio 5 days a week","Include omega-3 rich foods: salmon, walnuts, flaxseed","Discuss statin options with your doctor"]}
    if any(k in fn for k in ["thyroid","tsh","t3","t4"]):
        return {"reportType":"Thyroid Function Test","values":{"TSH":{"val":6.8,"unit":"mIU/L","min":0.4,"max":4.0,"status":"high"},"Free T4":{"val":0.7,"unit":"ng/dL","min":0.8,"max":1.8,"status":"low"},"Free T3":{"val":2.4,"unit":"pg/mL","min":2.3,"max":4.2,"status":"normal"}},"summary":"Thyroid panel indicates subclinical hypothyroidism with elevated TSH and borderline low Free T4. Clinical correlation with symptoms like fatigue and weight gain is recommended.","explanation":"Your thyroid is like a thermostat controlling your body's energy. TSH is the signal telling the thermostat to work harder — yours is high, meaning your thyroid isn't producing enough hormones. Think of it like turning up the heat because the room isn't warm enough.","tips":["Get a thyroid ultrasound as recommended","Ask your doctor about levothyroxine (thyroid hormone)","Avoid excess iodine supplements","Eat selenium-rich foods: Brazil nuts, eggs","Track symptoms: fatigue, weight, mood, cold sensitivity"]}
    if any(k in fn for k in ["glucose","sugar","diab","hba1c"]):
        return {"reportType":"Blood Glucose & HbA1c","values":{"Fasting Glucose":{"val":118,"unit":"mg/dL","min":70,"max":100,"status":"high"},"HbA1c":{"val":6.1,"unit":"%","min":0,"max":5.7,"status":"high"},"Insulin":{"val":14,"unit":"µU/mL","min":2,"max":25,"status":"normal"}},"summary":"Results indicate pre-diabetic range with fasting glucose of 118 mg/dL and HbA1c of 6.1%. Immediate lifestyle intervention is strongly advised to prevent progression to Type 2 diabetes.","explanation":"Think of insulin as a key that unlocks your cells to let sugar in for energy. Right now, your cells aren't responding as well to that key — sugar is building up in your blood. The good news: you are at a stage where diet and exercise changes can genuinely reverse this.","tips":["Cut refined carbs: white rice, bread, sugary drinks","Walk 20-30 min after each meal to lower blood sugar","Eat protein and fiber first at each meal","Target 7-8 hours of sleep every night","Retest HbA1c in 3 months to track progress"]}
    if any(k in fn for k in ["liver","lft","alt","ast","sgpt","sgot"]):
        return {"reportType":"Liver Function Test","values":{"ALT (SGPT)":{"val":68,"unit":"U/L","min":7,"max":40,"status":"high"},"AST (SGOT)":{"val":54,"unit":"U/L","min":10,"max":40,"status":"high"},"Alkaline Phosphatase":{"val":112,"unit":"U/L","min":44,"max":147,"status":"normal"},"Total Bilirubin":{"val":1.1,"unit":"mg/dL","min":0.2,"max":1.2,"status":"normal"},"Albumin":{"val":4.0,"unit":"g/dL","min":3.5,"max":5.0,"status":"normal"}},"summary":"Liver enzymes ALT and AST are elevated above normal range, suggesting hepatocellular stress or early inflammation. Further evaluation including ultrasound and viral hepatitis panel is warranted.","explanation":"Your liver is your body's main filter — it cleans toxins from your blood. ALT and AST are enzymes that leak out when liver cells are stressed, like water seeping through cracks in a dam. Your levels are elevated, but the overall liver structure appears intact.","tips":["Avoid alcohol completely until levels normalize","Stop herbal supplements or non-prescribed medications temporarily","Drink more water to support detoxification","Eat a low-fat, high-vegetable diet","Get an abdominal ultrasound and hepatitis B/C panel"]}
    # Default CBC
    return {"reportType":"Complete Blood Count (CBC)","values":{"Hemoglobin":{"val":11.2,"unit":"g/dL","min":13.5,"max":17.5,"status":"low"},"WBC":{"val":9800,"unit":"cells/µL","min":4500,"max":11000,"status":"normal"},"Platelets":{"val":210000,"unit":"/µL","min":150000,"max":400000,"status":"normal"},"RBC":{"val":3.9,"unit":"million/µL","min":4.5,"max":5.9,"status":"low"},"Hematocrit":{"val":34,"unit":"%","min":41,"max":53,"status":"low"},"MCV":{"val":72,"unit":"fL","min":80,"max":100,"status":"low"}},"summary":"CBC reveals microcytic anemia with low hemoglobin consistent with iron-deficiency anemia. Iron studies and dietary assessment are recommended.","explanation":"Think of red blood cells as delivery trucks carrying oxygen around your body. Yours are fewer and smaller than ideal — like having fewer, smaller trucks trying to deliver the same load. This is the most common type of anemia and is very treatable.","tips":["Increase iron-rich foods: red meat, spinach, lentils","Take iron supplements as directed (with Vitamin C for better absorption)","Avoid tea/coffee with meals — they block iron absorption","Get iron studies (serum ferritin) to confirm iron deficiency","Cook in cast iron pans — they add small amounts of dietary iron"]}

def analyze_report(file_path: str, file_name: str) -> dict:
    text = extract_text(file_path)
    if _openai_ok and settings.OPENAI_API_KEY:
        try:
            prompt = f"Analyze this lab report:\n\n{text[:4000]}" if len(text) > 100 else f"Generate analysis for file named: {file_name}"
            return call_openai(prompt)
        except Exception as e:
            print(f"OpenAI error: {e}")
    return demo_analysis(file_name)
