backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
openai==1.30.1
Pillow==10.3.0
pytesseract==0.3.10
PyMuPDF==1.24.3
